Projectile Stuff

+++ TK-Server/wServer/core/net/handlers/OtherHitHandler.cs
@@ -11,7 +11,7 @@ namespace wServer.core.net.handlers
         public override void Handle(Client client, NReader rdr, ref TickTime tickTime)
         {
            var Time = rdr.ReadInt32();
-           var BulletId = rdr.ReadByte();
+           var BulletId = rdr.ReadInt32();
            var ObjectId = rdr.ReadInt32();
            var TargetId = rdr.ReadInt32();
