+++ TK-Server/wServer/core/net/handlers/SquareHitHandler.cs
@@ -7,7 +7,7 @@ namespace wServer.core.net.handlers
     public class SquareHitHandler : IMessageHandler
     {
         public int Time { get; set; }
-        public byte BulletId { get; set; }
+        public int BulletId { get; set; }
         public int ObjectId { get; set; }
 
         public override MessageId MessageId => MessageId.SQUAREHIT;
@@ -15,7 +15,7 @@ namespace wServer.core.net.handlers
         public override void Handle(Client client, NReader rdr, ref TickTime tickTime)
         {
             Time = rdr.ReadInt32();
-            BulletId = rdr.ReadByte();
+            BulletId = rdr.ReadInt32();
             ObjectId = rdr.ReadInt32();
         }
     }
